split and added list_utils to utils dir. Moved deprecated list utils to a test where they were still used
